Changed move to match "remove all" call
Changed the mod's move to match the game's Remove All calls
Not sure if this will actually impact anything.  There were
warnings in the Unity Log about slots already containing references
but it turns out that vanilla does it as well.

The code doesn't seem to change anything, but it should be a more
correct call.